# docs/openapi/vanstro-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VanStro API",
    "version": "1.0.0",
    "description": "Generated method and path inventory from registered Hono routes. Request and response schemas remain contract-review requirements."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/analytics/pageviews": {
      "post": {
        "summary": "POST /analytics/pageviews",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  }
}